factored has_parts_mixin into compound
