+++ .gitignore
@@ -40,7 +40,7 @@ output/*.html
 output/*/index.html
 
 # Sphinx
-#docs/_build
+docs/_build
 
 # PyCharm
 .idea
